+++ catalog/controller/extension/payment/nuvei.php
@@ -506,11 +506,13 @@ class ControllerExtensionPaymentNuvei extends Controller
         $order_total    = $this->get_price($this->order_info['total']);
         
         $comment_details = '<br/>' 
-            . $this->language->get('Transaction ID: ') . $trans_id . '<br/>'
-            . $this->language->get('Related Transaction ID: ') . $rel_tr_id . '<br/>'
             . $this->language->get('Status: ') . $status . '<br/>'
             . $this->language->get('Transaction Type: ') . $transactionType . '<br/>'
-            . $this->language->get('Payment Method: ') . $payment_method . '<br/>';
+            . $this->language->get('Transaction ID: ') . $trans_id . '<br/>'
+            . $this->language->get('Related Transaction ID: ') . $rel_tr_id . '<br/>'
+            . $this->language->get('Payment Method: ') . $payment_method . '<br/>'
+            . $this->language->get('Total Amount: ') . $total_amount . '<br/>'
+            . $this->language->get('Currency: ') . NUVEI_CLASS::get_param('currency') . '<br/>';
         
         switch($status) {
             case 'CANCELED':