+++ system/library/nuvei/NUVEI_CLASS.php
@@ -8,7 +8,7 @@
  * @author Nuvei
  */
 
-const NUVEI_PLUGIN_V            = '1.8';
+const NUVEI_PLUGIN_V            = '1.9';
 const NUVEI_PLUGIN_CODE         = 'nuvei';
 const NUVEI_PLUGIN_TITLE        = 'Nuvei';