+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * Check if Order total and currency are correct when DMN come.
     * Added plugin version into plugin settings Help Tools.
     * Disable DCC when Order total is Zero.
+    * * Added the amount and the currency into Order's notes.
 ```
 
 # 1.7